requirements.txt
click==8.1.7
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0
pytz==2024.1
six==1.16.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
